Rename Xcode project
This renames the Xcode project from "29" to "TwentyNine".

Previously, the project exhibited errors from just the base
configuration that prevented it from running a simulation and building
properly. This was due to the Xcode project name starting with a number.

Now, the Xcode project is renamed as "TwentyNine". All errors have
receeded, and the simulation/project build without issues.