Set SuitImage to scale to fit
This changes the SuitImage scaling from filling to fitting the frame.

Previously, the SuitImage was set to fill the frame with a custom width
and height. This ended up being less adaptable when viewing on different
devices.

Now, the SuitImage's scaling is set to fit the frame. This allows it to
better dynamically adjust to screen sizes and orientations.