Remove padding from SuitImage
This eliminates the padding modifier from the struct.

Previously, a SuitImage contained a default padding. This ended up
causing it to appear too small when used in a CornerPip, or when the
background color was the same as its ancestor containers.

Now, a SuitImage contains no default padding. This allows it to expand
to to the full container size when needed.