Fix Deck/Hand not registering CardButton Cards
This resoloves an issue where a Card failed to update the Deck/Hand.

Previously, when a CardButton was clicked the environment Deck and Hand
failed to add/remove the respective Card, which caused the Score in the
MessageBox to never display.

Now, the models are updated with the Card when applicable, and the
Score/action buttons appear in the MessageBox when a Hand is full.